Update for v1803
Fixed a value, added a service , more apps to the blacklist and a better app uninstaller thanks to Neon Barron

+++ scripts/apps.ps1
@@ -8,43 +8,58 @@ $apps = @(
     "Microsoft.BingFinance"
     "Microsoft.BingNews"
     "Microsoft.BingSports"
+    "Microsoft.BingTranslator"
     "Microsoft.BingWeather"
     "Microsoft.FreshPaint"
-    "Microsoft.Getstarted"
+    "Microsoft.Microsoft3DViewer"
     "Microsoft.MicrosoftOfficeHub"
     "Microsoft.MicrosoftSolitaireCollection"
+    "Microsoft.MicrosoftPowerBIForWindows"
+    "Microsoft.MinecraftUWP"
     #"Microsoft.MicrosoftStickyNotes"
+    "Microsoft.NetworkSpeedTest"
     "Microsoft.Office.OneNote"
-    "Microsoft.OneConnect"
     "Microsoft.People"
+    "Microsoft.Print3D"
     "Microsoft.SkypeApp"
+    "Microsoft.Wallet"
     #"Microsoft.Windows.Photos"
     "Microsoft.WindowsAlarms"
     #"Microsoft.WindowsCalculator"
     "Microsoft.WindowsCamera"
+    "microsoft.windowscommunicationsapps"
     "Microsoft.WindowsMaps"
     "Microsoft.WindowsPhone"
     "Microsoft.WindowsSoundRecorder"
     #"Microsoft.WindowsStore"
-    "Microsoft.XboxApp"
+    #"Microsoft.XboxApp"
+    "Microsoft.XboxGameOverlay"
+    "Microsoft.XboxIdentityProvider"
+    "Microsoft.XboxSpeechToTextOverlay"
     "Microsoft.ZuneMusic"
     "Microsoft.ZuneVideo"
-    "microsoft.windowscommunicationsapps"
-    "Microsoft.MinecraftUWP"
-
+    
+    
     # Threshold 2 apps
     "Microsoft.CommsPhone"
     "Microsoft.ConnectivityStore"
+    "Microsoft.GetHelp"
+    "Microsoft.Getstarted"
     "Microsoft.Messaging"
     "Microsoft.Office.Sway"
     "Microsoft.OneConnect"
     "Microsoft.WindowsFeedbackHub"
+
+    # Creators Update apps
+    "Microsoft.Microsoft3DViewer"
+    #"Microsoft.MSPaint"
 
     #Redstone apps
     "Microsoft.BingFoodAndDrink"
     "Microsoft.BingTravel"
     "Microsoft.BingHealthAndFitness"
     "Microsoft.WindowsReadingList"
+
 
     # non-Microsoft
     "9E2F88E3.Twitter"
@@ -67,6 +82,18 @@ $apps = @(
     "Facebook.Facebook"
     "flaregamesGmbH.RoyalRevolt2"
     "Playtika.CaesarsSlotsFreeCasino"
+    "A278AB0D.MarchofEmpires"
+    "KeeperSecurityInc.Keeper"
+    "ThumbmunkeysLtd.PhototasticCollage"
+    "XINGAG.XING"
+    "89006A2E.AutodeskSketchBook"
+    "D5EA27B7.Duolingo-LearnLanguagesforFree"
+    "46928bounde.EclipseManager"
+    "ActiproSoftwareLLC.562882FEEB491" # next one is for the Code Writer from Actipro Software LLC
+    "DolbyLaboratories.DolbyAccess"
+    "SpotifyAB.SpotifyMusic"
+    "A278AB0D.DisneyMagicKingdoms"
+    "WinZipComputing.WinZipUniversal"
 
     # apps which cannot be removed using Remove-AppxPackage
     #"Microsoft.BioEnrollment"
@@ -76,11 +103,18 @@ $apps = @(
     #"Microsoft.XboxGameCallableUI"
     #"Microsoft.XboxIdentityProvider"
     #"Windows.ContactSupport"
+
+    #Added by W10 Cleanser
+    "Microsoft.Advertising.Xaml"
 )
 
 foreach ($app in $apps) {
-    Get-AppxPackage -Name $app -AllUsers | Remove-AppxPackage -AllUsers
-    Get-AppXProvisionedPackage -Online |
-        where DisplayName -EQ $app |
-        Remove-AppxProvisionedPackage -Online
+    $appInstalled = Get-AppxPackage $app | % {$_.name -eq $app}
+    if ($appInstalled -eq $true){
+        Write-Output "Removing: $app"
+        Get-AppxPackage -Name $app -AllUsers | Remove-AppxPackage -AllUsers
+        Get-AppXProvisionedPackage -Online |
+            Where-Object DisplayName -EQ $app |
+            Remove-AppxProvisionedPackage -Online
+        }
 }
